+++ packages/search/global-search/src/components/GlobalQuickSearchWrapper.tsx
@@ -46,10 +46,10 @@ export type AdvancedSearchEvent = {
    */
   category: String;
   /**
-   * original orignial event, this is useful when need to check if the click was to open in new tab or in same tab
+   * orignial event, this is useful when need to check if the click was to open in new tab or in same tab
    * but if consumer wanna cancel the event {@link preventDefault} should be used
    */
-  orignialEvent: Object;
+  originalEvent: Object;
 };
 export interface Props {
   /**
@@ -204,7 +204,7 @@ export default class GlobalQuickSearchWrapper extends React.Component<Props> {
         preventDefault: () => (preventEventDefault = true),
         query, // query entered by the user
         category: entity,
-        orignialEvent: e,
+        originalEvent: e,
       });
 
       if (preventEventDefault) {